Revert "sdp.py / W value members to integers"
This reverts commit b830ceb14662e6603c1217f2a2ab65a0aa9e8ba4.

+++ sdp.py
@@ -39,8 +39,7 @@ class SDP:
             elif key[-1] == 'W':
                 if not isinstance(val, str):
                     raise Exception('List of Values _MUST_BE_ string of numbers')
-                #self.add_value(key[:-1], list(map(float, val.split(' '))))
-                self.add_value(key[:-1], list(map(int, val.split(' ')))) # let the value members to be integers
+                self.add_value(key[:-1], list(map(float, val.split(' '))))
             else:
                 if not isinstance(val, str):
                     raise Exception('Data _MUST_BE_ string')
